real_estate_bot/requirements.txt
aiogram==3.7.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0